ai_studio/app/pyproject.toml
[project]
name = "gemini-api"
version = "0.1.0"
description = "Gemini API FastAPI Service"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "google-genai>=0.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"